split gem service into multiple files
